+++ examples/main/main.tf
@@ -1,9 +1,13 @@
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = ">= 1.0.0"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.56"
+      version = ">= 2.0.0"
     }
   }
+}
+
+provider "azurerm" {
+  features {}
 }