+++ locals.tf
@@ -2,25 +2,23 @@ locals {
   name_prefix  = var.name_prefix != "" ? replace(var.name_prefix, "/[a-z0-9]$/", "$0-") : ""
   default_name = lower("${local.name_prefix}${var.stack}-${var.client_name}-${var.location_short}-${var.environment}")
 
-  appgw_name = var.appgw_name != "" ? var.appgw_name : join("-", [local.default_name, "appgw"])
+  appgw_name = coalesce(var.custom_appgw_name, join("-", [local.default_name, "appgw"]))
 
-  subnet_name = var.custom_subnet_name != "" ? [var.custom_subnet_name] : [join("-", [local.default_name, "subnet"])]
-  subnet_id   = var.create_subnet ? module.azure_network_subnet.subnet_ids[0] : var.subnet_id
+  subnet_name = coalesce(var.custom_subnet_name, join("-", [local.default_name, "subnet"]))
+  subnet_id   = var.create_subnet ? module.azure_network_subnet["appgw_subnet"].subnet_id : var.subnet_id
 
-  nsg_ids = var.create_nsg ? {
-    element(local.subnet_name, 0) = join("", module.azure_network_security_group.network_security_group_id)
-  } : {}
+  nsg_name = var.create_nsg ? module.azure_network_security_group["appgw_nsg"].network_security_group_name : null
 
   nsr_https_name       = coalesce(var.custom_nsr_https_name, join("-", [local.default_name, "https-nsr"]))
   nsr_healthcheck_name = coalesce(var.custom_nsr_healthcheck_name, join("-", [local.default_name, "appgw-healthcheck-nsr"]))
 
-  ip_name  = var.ip_name != "" ? var.ip_name : join("-", [local.default_name, "pubip"])
-  ip_label = var.ip_label != "" ? var.ip_label : join("-", [local.default_name, "pubip"])
+  ip_name  = coalesce(var.custom_ip_name, join("-", [local.default_name, "pubip"]))
+  ip_label = coalesce(var.custom_ip_label, join("-", [local.default_name, "pubip"]))
 
-  frontend_ip_configuration_name      = var.frontend_ip_configuration_name != "" ? var.frontend_ip_configuration_name : join("-", [local.default_name, "frontipconfig"])
-  frontend_priv_ip_configuration_name = var.frontend_priv_ip_configuration_name != "" ? var.frontend_ip_configuration_name : join("-", [local.default_name, "frontipconfig-priv"])
+  frontend_ip_configuration_name      = coalesce(var.custom_frontend_ip_configuration_name, join("-", [local.default_name, "frontipconfig"]))
+  frontend_priv_ip_configuration_name = coalesce(var.custom_frontend_priv_ip_configuration_name, join("-", [local.default_name, "frontipconfig-priv"]))
 
-  gateway_ip_configuration_name = var.gateway_ip_configuration_name != "" ? var.gateway_ip_configuration_name : join("-", [local.default_name, "gwipconfig"])
+  gateway_ip_configuration_name = coalesce(var.custom_gateway_ip_configuration_name, join("-", [local.default_name, "gwipconfig"]))
 
   enable_waf = var.sku == "WAF_v2" ? true : false