+++ r-network.tf
@@ -1,36 +1,40 @@
 module "azure_network_subnet" {
   source  = "claranet/subnet/azurerm"
-  version = "3.0.0"
+  version = "4.2.1"
+
+  for_each = var.create_subnet ? toset(["appgw_subnet"]) : []
 
   environment         = var.environment
   location_short      = var.location_short
   client_name         = var.client_name
   stack               = var.stack
-  resource_group_name = var.subnet_resource_group_name != "" ? var.subnet_resource_group_name : var.resource_group_name
+  resource_group_name = coalesce(var.subnet_resource_group_name, var.resource_group_name)
 
   virtual_network_name = var.virtual_network_name
 
-  custom_subnet_names = var.create_subnet ? local.subnet_name : []
-  subnet_cidr_list    = var.create_subnet ? [var.subnet_cidr] : []
+  custom_subnet_name = local.subnet_name
+  subnet_cidr_list   = [var.subnet_cidr]
 
-  network_security_group_ids = local.nsg_ids
+  network_security_group_name = local.nsg_name
 
-  route_table_ids = var.create_subnet ? var.route_table_ids : {}
+  route_table_name = var.route_table_name
+  route_table_rg   = var.route_table_rg
 }
 
 module "azure_network_security_group" {
   source  = "claranet/nsg/azurerm"
-  version = "3.0.0"
+  version = "4.1.1"
+
+  for_each = var.create_nsg ? toset(["appgw_nsg"]) : []
 
   client_name         = var.client_name
   environment         = var.environment
   stack               = var.stack
-  resource_group_name = var.subnet_resource_group_name != "" ? var.subnet_resource_group_name : var.resource_group_name
+  resource_group_name = coalesce(var.subnet_resource_group_name, var.resource_group_name)
   location            = var.location
   location_short      = var.location_short
 
-  custom_network_security_group_names = var.custom_nsg_name != [""] ? [var.custom_nsg_name] : [""]
-  network_security_group_instances    = var.create_nsg ? 1 : 0
+  custom_network_security_group_name = var.custom_nsg_name
 
   extra_tags = merge(local.default_tags, var.extra_tags)
 }
@@ -40,8 +44,8 @@ resource "azurerm_network_security_rule" "web" {
 
   name = local.nsr_https_name
 
-  resource_group_name         = var.subnet_resource_group_name != "" ? var.subnet_resource_group_name : var.resource_group_name
-  network_security_group_name = module.azure_network_security_group.network_security_group_name[0]
+  resource_group_name         = coalesce(var.subnet_resource_group_name, var.resource_group_name)
+  network_security_group_name = module.azure_network_security_group["appgw_nsg"].network_security_group_name
 
   priority  = 100
   direction = "Inbound"
@@ -61,8 +65,8 @@ resource "azurerm_network_security_rule" "allow_health_probe_app_gateway" {
 
   name = local.nsr_healthcheck_name
 
-  resource_group_name         = var.subnet_resource_group_name != "" ? var.subnet_resource_group_name : var.resource_group_name
-  network_security_group_name = module.azure_network_security_group.network_security_group_name[0]
+  resource_group_name         = coalesce(var.subnet_resource_group_name, var.resource_group_name)
+  network_security_group_name = module.azure_network_security_group["appgw_nsg"].network_security_group_name
 
   priority  = 101
   direction = "Inbound"